requirements.txt
python-dotenv==1.0.0
langchain-community
langchain-openai
langchain==0.1.11
openai==1.9.0
google-generativeai==0.3.1
langchain-google-genai